requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contextlib2==21.6.0
cryptography==44.0.3
defusedxml==0.7.1
Django==5.2
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
injector==0.22.0
oauthlib==3.2.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
